Added nekoweb site files
Because of neocities' CSP policies, my online JavaScript tools need to be hosted on a different site: nekoweb. These are the html files on that site.
The RSS-Gen ones have the JS embedded for convenience, and it's been a while, but I believe that the Lane-Detector js file is hosted on neocities.